Removed Convert.ChangeType calls because of poor performance (DynamicReflectionDelegateFactory now uses IConvertible.ToTypeName, while ExpressionReflectionDelegateFactory uses Convert.ToTypeName). Method binder moved to a separate file and documented, + code style fixes.
