+++ Src/Newtonsoft.Json/Utilities/TypeExtensions.cs
@@ -274,225 +274,6 @@ namespace Newtonsoft.Json.Utilities
 
 #if !DOTNET
         
-        private static readonly Type[] PrimitiveTypes = new Type[] 
-        {
-            typeof(bool),  typeof(char),   typeof(sbyte), typeof(byte),
-            typeof(short), typeof(ushort), typeof(int),   typeof(uint),
-            typeof(long),  typeof(ulong),  typeof(float), typeof(double)
-        };
-
-        private static readonly int[] WideningMasks = new int[]
-        {
-            0x0001,        0x0FE2,         0x0D54,        0x0FFA,
-            0x0D50,        0x0FE2,         0x0D40,        0x0F80,
-            0x0D00,        0x0E00,         0x0C00,        0x0800
-        };
-        
-        internal static bool CanConvertPrimitive(Type from, Type to)
-        {
-            if (from == to) return true;
-
-            int fromMask = 0, toMask = 0;
-
-            for (int i = 0; i < PrimitiveTypes.Length && (fromMask == 0 || toMask == 0); i++)
-            {
-                if (PrimitiveTypes[i] == from)
-                {
-                    fromMask = WideningMasks[i];
-                }
-                else if (PrimitiveTypes[i] == to)
-                {
-                    toMask = 1 << i;
-                }
-            }
-            
-            return (fromMask & toMask) != 0;
-        }
-
-        private static bool FilterParameters(ParameterInfo[] parameters, IList<Type> types, bool enableParamArray)
-        {
-            if (parameters == null)
-                throw new ArgumentNullException(nameof(parameters));
-            if (types == null)
-                throw new ArgumentNullException(nameof(types));
-
-            if (parameters.Length == 0)
-            {
-                // fast check for parameterless methods
-                return types.Count == 0;
-            }
-            
-            if (parameters.Length > types.Count)
-            {
-                // not all declared parameters were specified (optional parameters are not supported)
-                return false;
-            }
-
-            // check if the last parameter is ParamArray
-            Type paramArrayType = null;
-            
-            if (enableParamArray &&
-                parameters[parameters.Length - 1].ParameterType.IsArray &&
-                parameters[parameters.Length - 1].IsDefined(typeof(ParamArrayAttribute)))
-            {
-                paramArrayType = parameters[parameters.Length - 1].ParameterType.GetElementType();
-            }
-
-            if (paramArrayType == null && parameters.Length != types.Count)
-            {
-                // when there's no ParamArray, number of parameters should match
-                return false;
-            }
-            
-            for (int i = 0; i < types.Count; i++)
-            {
-                var paramType = (paramArrayType != null && i >= parameters.Length - 1) ? paramArrayType : parameters[i].ParameterType;
-
-                // exact match with provided type
-                if (paramType == types[i]) continue;
-
-                // parameter of type object matches anything
-                if (paramType == typeof(object)) continue;
-
-                if (paramType.IsPrimitive())
-                {
-                    // primitive parameter can only be assigned from compatible primitive type
-                    if (!types[i].IsPrimitive() || !CanConvertPrimitive(types[i], paramType))
-                        return false;
-                }
-                else
-                {
-                    if (!paramType.IsAssignableFrom(types[i]))
-                        return false;
-                }
-            }
-
-            return true;
-        }
-
-        private class ParametersMatchComparer : IComparer<ParameterInfo[]>
-        {
-            private readonly IList<Type> _types;
-            private readonly bool _enableParamArray;
-
-            public ParametersMatchComparer(IList<Type> types, bool enableParamArray)
-            {
-                if (types == null)
-                    throw new ArgumentNullException(nameof(types));
-
-                _types = types;
-                _enableParamArray = enableParamArray;
-            }
-
-            public int Compare(ParameterInfo[] parameters1, ParameterInfo[] parameters2)
-            {
-                if (parameters1 == null)
-                    throw new ArgumentNullException(nameof(parameters1));
-                if (parameters2 == null)
-                    throw new ArgumentNullException(nameof(parameters2));
-
-                // parameterless method wins
-                if (parameters1.Length == 0) return -1;
-                if (parameters2.Length == 0) return 1;
-
-                Type paramArrayType1 = null, paramArrayType2 = null;
-
-                if (_enableParamArray)
-                {
-                    if (parameters1[parameters1.Length - 1].ParameterType.IsArray &&
-                        parameters1[parameters1.Length - 1].IsDefined(typeof(ParamArrayAttribute)))
-                    {
-                        paramArrayType1 = parameters1[parameters1.Length - 1].ParameterType.GetElementType();
-                    }
-
-                    if (parameters2[parameters2.Length - 1].ParameterType.IsArray &&
-                        parameters2[parameters2.Length - 1].IsDefined(typeof(ParamArrayAttribute)))
-                    {
-                        paramArrayType2 = parameters2[parameters2.Length - 1].ParameterType.GetElementType();
-                    }
-
-                    // A method using params always loses to one not using params
-                    if (paramArrayType1 != null && paramArrayType2 == null) return 1;
-                    if (paramArrayType2 != null && paramArrayType1 == null) return -1;
-                }
-
-                for (int i = 0; i < _types.Count; i++)
-                {
-                    var type1 = (paramArrayType1 != null && i >= parameters1.Length - 1) ? paramArrayType1 : parameters1[i].ParameterType;
-                    var type2 = (paramArrayType2 != null && i >= parameters2.Length - 1) ? paramArrayType2 : parameters2[i].ParameterType;
-
-                    // exact match between parameter types doesn't change score
-                    if (type1 == type2) continue;
-
-                    // exact match with source type decides winner immediately
-                    if (type1 == _types[i]) return -1;
-                    if (type2 == _types[i]) return 1;
-
-                    var r = ChooseMorePreciseType(type1, type2);
-                    if (r != 0) return r;
-                }
-
-                return 0;
-            }
-
-            private static int ChooseMorePreciseType(Type type1, Type type2)
-            {
-                if (type1.IsByRef || type2.IsByRef)
-                {
-                    if (type1.IsByRef && type2.IsByRef)
-                    {
-                        type1 = type1.GetElementType();
-                        type2 = type2.GetElementType();
-                    }
-                    else if (type1.IsByRef)
-                    {
-                        type1 = type1.GetElementType();
-                        if (type1 == type2) return 1;
-                    }
-                    else
-                    {
-                        type2 = type2.GetElementType();
-                        if (type2 == type1) return -1;
-                    }
-                }
-
-                bool c1FromC2, c2FromC1;
-
-                if (type1.IsPrimitive() && type2.IsPrimitive())
-                {
-                    c1FromC2 = CanConvertPrimitive(type2, type1);
-                    c2FromC1 = CanConvertPrimitive(type1, type2);
-                }
-                else
-                {
-                    c1FromC2 = type1.IsAssignableFrom(type2);
-                    c2FromC1 = type2.IsAssignableFrom(type1);
-                }
-
-                if (c1FromC2 == c2FromC1) return 0;
-
-                return c1FromC2 ? 1 : -1;
-            }
-
-        }
-        
-        private static MethodBase SelectMethod(IEnumerable<MethodBase> candidates, BindingFlags bindingFlags, IList<Type> types)
-        {
-            if (candidates == null)
-                throw new ArgumentNullException(nameof(candidates));
-            if (types == null)
-                throw new ArgumentNullException(nameof(types));
-
-            // ParamArrays are not supported by ReflectionDelegateFactory
-            // They will be treated like ordinary array arguments
-            const bool enableParamArray = false;
-
-            return candidates
-                .Where(m => TestAccessibility(m, bindingFlags) && FilterParameters(m.GetParameters(), types, enableParamArray))
-                .OrderBy(m => m.GetParameters(), new ParametersMatchComparer(types, enableParamArray))
-                .FirstOrDefault();
-        }
-
         public static MethodInfo GetMethod(this Type type, string name)
         {
             return type.GetMethod(name, DefaultFlags);
@@ -515,7 +296,7 @@ namespace Newtonsoft.Json.Utilities
 
         public static MethodInfo GetMethod(this Type type, string name, BindingFlags bindingFlags, object placeHolder1, IList<Type> parameterTypes, object placeHolder2)
         {
-            return (MethodInfo)SelectMethod(type.GetTypeInfo().DeclaredMethods.Where(m => name == null || m.Name == name), bindingFlags, parameterTypes);
+            return MethodBinder.SelectMethod(type.GetTypeInfo().DeclaredMethods.Where(m => (name == null || m.Name == name) && TestAccessibility(m, bindingFlags)), parameterTypes);
         }
 
         public static IEnumerable<ConstructorInfo> GetConstructors(this Type type)
@@ -535,7 +316,7 @@ namespace Newtonsoft.Json.Utilities
 
         public static ConstructorInfo GetConstructor(this Type type, BindingFlags bindingFlags, object placeholder1, IList<Type> parameterTypes, object placeholder2)
         {
-            return (ConstructorInfo)SelectMethod(type.GetTypeInfo().DeclaredConstructors, bindingFlags, parameterTypes);
+            return MethodBinder.SelectMethod(type.GetConstructors(bindingFlags), parameterTypes);
         }
         
         public static MemberInfo[] GetMember(this Type type, string member)
